referral_system/requirements.txt
asgiref==3.7.2
Babel==2.12.1
Django==3.2.20
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
flake8==5.0.4
gunicorn==20.1.0
importlib-metadata==4.2.0
inflection==0.5.1
isort==5.11.5
mccabe==0.7.0
packaging==23.1
pep8==1.7.1
phonenumbers==8.13.18
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.8.0
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
zipp==3.15.0
